test(screamingface-engine): name the board in every HealthBench identity test
Four test names existed identically in test_healthbench_definition.py and
test_healthbench_professional.py — `pytest -k test_the_exam_is_registered_under
_its_id` matched two boards and a failing report named neither.

The two files also still mirrored the old source split. They now mirror the
merged definition.py: shared assertions once (grader template, judge pinning,
retrieval-free Candidate, expression size — the last three looped over both
boards so a third cannot ship a different judge), then one section per board
with board-named tests.

test_healthbench_aggregate.py is deliberately untouched: its tests pass a
synthetic benchmark_id and exercise the shared reducer, not either board.

Refs: OME-903
